+++ syllabus/1-Cryptography/6-Hash_Based_Data_Structures/Activity-merkle-tree/src/main.rs
