--- syllabus/1-Cryptography/Activity-Many-Time-Pad/Many_Time_Pad_Activity.md
@@ -1,63 +1,0 @@
-# Activity: Crack the Many Time Pad
-
-## Citation
-
-This activity is cribbed from Dan Boneh's [Coursera Cryptography I](https://www.coursera.org/learn/crypto/quiz/KZ9js/week-1-programming-assignment-optional/attempt) course.
-
-The goal of this activity is to illustrate that while the one-time pad is uncrackable if it is really only used once, reusing the key multiple times makes the cipher trivially insecure.
-
-## Introduction
-
-The symmetric one-time pad is known to be secure when the key is only used once.
-In practice key distribution is not always practical, and users sometimes make the critical mistake of reusing a pre-shared key.
-
-In this activity, students will experience first hand why reusing the key is detrimental to security.
-
-## The Challenge
-
-The following several ciphertexts were all intercepted.
-We believe they may have been encrypted with the same key.
-Your task is to extract _all_ of the plaintexts as well as the original key.
-
-```text
-15481555260c011c535055565d671501024255554b00350d131d0a0a024c0014000a024501061b0105490c464312160b4f190b550e05490d03191b5746001c4c0b0f0b4b034f
-064f0610000b011c4e041645135f114f4a1f0a015241024511061d1b074e08464f030a45171c174f060c024410411c0e00164f16090a54130d00030e000c01021d1c0a4c1c041d4d0b0b00044f06045342094918004e3a41141a060011584c1b5512521f18520a4e351220411c110d571b1d4b124e031d03005267191c000d0319150006010d4e334f0648010a100a48540e00040c4e09010f0d080247491a1b110006551b1e0140
-034904161d0c0259490345020048151b0d1116451300100a001e490008000b0f470c18040f541f00000c1a0c4303061c001e1b064c174313051c1b1e4e084e000800025511061c4d0c1d54074f1b4f5707044b530301060005070b45081d021d0009144f1e451d070a55034113100f411a0c45054e0e0418091b43161d1d0701065454084f0b0b474201011e1b55080600154f1647
-2f001f0716001e59541f450d135b114f4c50010011451811001205061d450b46440418040115010a42491a4f16411d0d45134f01034448001a094f04450c1b1e001a1c0e5028174d0a1c101652541b4f420d4105004e0745071c000c170d40595909074f03450a0a45541f411a150f4554064e020b01000113175359
-005350061d06051c54094507174e1b02480345081d521345131d0d4f034f1d03000603081318171742490048060007014e104f0205084c4105024f1a4101174c1e0f1c5350031c0e0a03115350060047100053000c18114c1d491704101d090b00071c0b4d450e1d0c450241061b59441b4f410f0a4f1c09171645054900074f051b4c0e0c0c4e0852540d040a1b471d4e0545141a1a00000747
-0800121015040259541f4517174c18065e15450b175756151d001a060c49030f540c0916431b020a00000d474314034842121b0209014e4118040a5746060b000d1d454f1641302e314e151d445408410f0000070d0b1b521d4552040d104c0d48035206034519071141120d17540a4f1706410d4e0c1c090b1545571d1b48181d1d430f4f1d060e53541f1d1a1903484c04410247
-02520905060a0f0c5202000b1144111c0d1109091d571301521d060143431a15540a080c0218520a160a0b410d06164400000601040b55154c191c12521c4e0408180c4e17410d02451d1d144e541a50420a5253061c1141100c52040017030c4e120141
-0f4f04550b0a190b001b001c010154214204451c1d550445111c00011d0e
-```
-
-# Helpful Hints
-
-- XOR the ciphertexts together, and consider what happens when a space is XORed with a character in [a-zA-Z].
-- Consider Crib dragging: https://samwho.dev/blog/toying-with-cryptography-crib-dragging/
-
-Potential Solution: https://joshorndorff.github.io/DanBonehsCryptoExercises/crib-dragging/ is a tool used to perform crib dragging interactively
-
-## Instructions
-
-The intention of this activity is for it to be fun and engaging.
-These instructions are meant to serve as guidelines to help you complete the activity:
-
-- Team up with 2-4 students to complete this activity.
-
-- Write a program, preferably in Rust, that finds the key to generate the plaintext from the provided cipher texts (consider the hints).<br>
-  The general steps are:
-
-  1. Find the length of the longest input cipher text.
-  1. Generate a key of that length.
-  1. Find what the correct key is...
-
-- Don't hesitate to ask for support if you're feeling stuck, or just ask your peers.
-
-- Spend 1-2 hours to crack the Many Time Pad.
-
-- If you've cracked it before other teams have, find ways to improve your solution.<br>
-Some suggestions:
-  - Create a tool that automates the cipher key generation
-  - Add a way to generate new cipher texts
-  - Create your own cipher texts using other cipher methods
-  - Provide a new set of ciphertexts that were intentionally constructed not to use the most common English words.
