+++ syllabus/1-Cryptography/README.md
@@ -9,14 +9,16 @@ Personal Cryptography and How It's Used
 ### Day 1
 
 1. [Introduction to Cryptography](./1-Crypto_Introduction/lesson-plan.md)
+   1. [Activity: Many Time Pad](./1-Crypto_Introduction/Activity-Many-Time-Pad/)
 2. [Digital Signatures (Part 1)](./2-Digital_Signatures_Basic/lesson-plan.md)
 
 ### Day 2
 
-3. [Digital Signatures (Part 2)](./3-Digital_Signatures_Advanced/lesson-plan.md)
-4. [Other Exotic Primitives](./4-Exotic_Primitives/lesson-plan.md)
+1. [Digital Signatures (Part 2)](./3-Digital_Signatures_Advanced/lesson-plan.md)
+   1. [Activity: AES](./3-Digital_Signatures_Advanced/Activity-aes-modes/)
+1. [Other Exotic Primitives](./4-Exotic_Primitives/lesson-plan.md)
 
 ### Day 3
 
-5. [Hashes](./5-Hashes/lesson-plan.md)
-6. [Hash-Based Data Structures](./6-Hash_Based_Data_Structures/lesson-plan.md)
+1. [Hashes](./5-Hashes/lesson-plan.md)
+1. [Hash-Based Data Structures](./6-Hash_Based_Data_Structures/lesson-plan.md)
